requirements.txt
annotated-types==0.7.0
asgiref==3.11.0
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
cbor2==5.7.1
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
chess==1.11.2
constantly==23.10.4
contextlib2==21.6.0
cryptography==46.0.3
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
django-ninja==1.5.0
django-ninja-extra==0.30.6
django-ninja-jwt==5.4.2
dnspython==2.8.0
email-validator==2.3.0
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
injector==0.22.0
msgpack==1.1.2
pillow==12.0.0
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-dotenv==1.2.1
redis==7.1.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.11.0
zope.interface==8.1.1
